refactor: move outputdistribution enum to enums folder
- Move OutputDistribution.cs from src/Normalizers to src/Enums
- Update namespace from AiDotNet.Normalizers to AiDotNet.Enums
- Add using AiDotNet.Enums to NormalizationParameters.cs and QuantileTransformer.cs
- Update property type references to use unqualified OutputDistribution
