+++ src/Normalizers/QuantileTransformer.cs
@@ -1,3 +1,5 @@
+using AiDotNet.Enums;
+
 namespace AiDotNet.Normalizers;
 
 /// <summary>
